+++ pom.xml
@@ -147,7 +147,7 @@
         <configuration>
           <archive>
             <manifest>
-              <mainClass>org.opengis.cite.gpkg12.TestNGController</mainClass>
+              <mainClass>org.opengis.cite.gpkg12.nsg.TestNGController</mainClass>
             </manifest>
           </archive>
           <descriptors>
