+++ src/main/resources/org/opengis/cite/gpkg12/nsg/ets.properties
