+++ src/test/java/org/opengis/cite/gpkg12/nsg/VerifyTestNGController.java
@@ -1,4 +1,4 @@
-package org.opengis.cite.gpkg12;
+package org.opengis.cite.gpkg12.nsg;
 
 import static org.junit.Assert.*;
 
@@ -20,7 +20,6 @@ import net.sf.saxon.s9api.XdmValue;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.opengis.cite.gpkg12.TestNGController;
 import org.opengis.cite.gpkg12.TestRunArg;
 import org.opengis.cite.gpkg12.util.XMLUtils;
 import org.w3c.dom.Document;
